# MTLS_UI/src/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/assignments": {
            "post": {
                "operationId": "AssignmentController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAssignmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bài tập đã được tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu đầu vào không hợp lệ"
                    }
                },
                "summary": "Tạo bài tập mới",
                "tags": ["assignments"]
            },
            "get": {
                "operationId": "AssignmentController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách bài tập được giao"
                    }
                },
                "summary": "Lấy danh sách tất cả bài tập được giao",
                "tags": ["assignments"]
            }
        },
        "/assignments/{id}": {
            "get": {
                "operationId": "AssignmentController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thông tin bài tập"
                    },
                    "404": {
                        "description": "Không tìm thấy bài tập"
                    }
                },
                "summary": "Lấy thông tin bài tập theo ID",
                "tags": ["assignments"]
            },
            "patch": {
                "operationId": "AssignmentController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAssignmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bài tập đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy bài tập"
                    }
                },
                "summary": "Cập nhật bài tập",
                "tags": ["assignments"]
            },
            "delete": {
                "operationId": "AssignmentController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Bài tập đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy bài tập"
                    }
                },
                "summary": "Xóa bài tập",
                "tags": ["assignments"]
            }
        },
        "/assignments/section/{sectionId}": {
            "get": {
                "operationId": "AssignmentController_findBySection",
                "parameters": [{
                    "name": "sectionId",
                    "required": true,
                    "in": "path",
                    "description": "ID của section",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách bài tập của section"
                    }
                },
                "summary": "Lấy danh sách bài tập theo section",
                "tags": ["assignments"]
            }
        },
        "/assignments/{id}/submissions": {
            "get": {
                "operationId": "AssignmentController_findOneWithSubmissions",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về bài tập và danh sách bài nộp"
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp của bài tập",
                "tags": ["assignments"]
            }
        },
        "/assignments/{id}/students": {
            "get": {
                "operationId": "AssignmentController_getStudentsWhoSubmitted",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách học viên đã làm bài"
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách học viên đã làm bài tập",
                "tags": ["assignments"]
            }
        },
        "/assignments/{id}/stats": {
            "get": {
                "operationId": "AssignmentController_getAssignmentStats",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của bài tập",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thống kê của bài tập"
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy thống kê bài tập",
                "tags": ["assignments"]
            }
        },
...
        "/do-assignments/student/{studentId}/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_findByStudentAndAssignment",
                "parameters": [{
                        "name": "studentId",
                        "required": true,
                        "in": "path",
                        "description": "ID của học viên",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "assignmentId",
                        "required": true,
                        "in": "path",
                        "description": "ID của assignment",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp của học viên và assignment."
                    },
                    "404": {
                        "description": "Học viên hoặc assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp theo student ID và assignment ID",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_findByAssignment",
                "parameters": [{
                    "name": "assignmentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của assignment",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách bài nộp của assignment."
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách bài nộp theo assignment ID",
                "tags": ["do-assignments"]
            }
        },
...
        "/do-assignments/statistics/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_getAssignmentStatistics",
                "parameters": [{
                    "name": "assignmentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của assignment",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Thống kê điểm số."
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy thống kê điểm số cho một assignment",
                "tags": ["do-assignments"]
            }
        },
        "/do-assignments/students/assignment/{assignmentId}": {
            "get": {
                "operationId": "DoAssignmentController_getStudentSubmissionsByAssignment",
                "parameters": [{
                    "name": "assignmentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của assignment",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Danh sách học sinh và trạng thái chấm điểm."
                    },
                    "404": {
                        "description": "Assignment không tồn tại."
                    }
                },
                "summary": "Lấy danh sách học sinh đã làm assignment và trạng thái chấm điểm",
                "tags": ["do-assignments"]
            }
        },
...
    },
    "info": {
        "title": "API Documentation",
        "description": "The API's MTLS description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateAssignmentDto": {
                "type": "object",
                "properties": {
                    "section": {
                        "type": "string",
                        "description": "ID của section mà assignment thuộc về",
                        "example": "6507d6f6c31c7948c831a123"
                    },
                    "title": {
                        "type": "string",
                        "description": "Tiêu đề của assignment",
                        "example": "Bài tập về nhà: Lập trình hướng đối tượng"
                    },
                    "questionText": {
                        "type": "string",
                        "description": "Nội dung câu hỏi của assignment",
                        "example": "Hãy viết một chương trình minh họa tính kế thừa trong OOP"
                    },
                    "description": {
                        "type": "string",
                        "description": "Mô tả chi tiết về assignment",
                        "example": "Bài tập này giúp sinh viên hiểu rõ hơn về tính kế thừa trong lập trình hướng đối tượng"
                    },
                    "fileUrl": {
                        "type": "string",
                        "description": "URL của file đính kèm (nếu có)",
                        "example": "https://example.com/files/assignment1.pdf"
                    },
                    "duration": {
                        "type": "number",
                        "description": "Thời gian làm bài (tính bằng phút)",
                        "example": 60,
                        "default": 0
                    }
                },
                "required": ["section", "title"]
            },
            "UpdateAssignmentDto": {
                "type": "object",
                "properties": {
                    "section": {
                        "type": "string",
                        "description": "ID của section mà assignment thuộc về",
                        "example": "6507d6f6c31c7948c831a123"
                    },
                    "title": {
                        "type": "string",
                        "description": "Tiêu đề của assignment",
                        "example": "Bài tập về nhà: Lập trình hướng đối tượng (Cập nhật)"
                    },
                    "questionText": {
                        "type": "string",
                        "description": "Nội dung câu hỏi của assignment",
                        "example": "Hãy viết một chương trình minh họa tính kế thừa và đa hình trong OOP"
                    },
                    "description": {
                        "type": "string",
                        "description": "Mô tả chi tiết về assignment",
                        "example": "Bài tập này giúp sinh viên hiểu rõ hơn về tính kế thừa và đa hình trong lập trình hướng đối tượng"
                    },
                    "fileUrl": {
                        "type": "string",
                        "description": "URL của file đính kèm (nếu có)",
                        "example": "https://example.com/files/assignment1_updated.pdf"
                    },
                    "duration": {
                        "type": "number",
                        "description": "Thời gian làm bài (tính bằng phút)",
                        "example": 90
                    }
                }
            },
...
        }
    }
}